# lib/api-zod/src/generated/types/messageRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageRecordAudience } from './messageRecordAudience';
import type { MessageRecordChannel } from './messageRecordChannel';
import type { MessageRecordMetadata } from './messageRecordMetadata';

export interface MessageRecord {
  id: number;
  recipient: string;
  body: string;
  channel: MessageRecordChannel;
  audience: MessageRecordAudience;
  status: string;
  provider: string;
  /** @nullable */
  recipientPhone?: string | null;
  /** @nullable */
  recipientName?: string | null;
  /** @nullable */
  providerMessageId?: string | null;
  /** @nullable */
  sentAt?: Date | null;
  /** @nullable */
  deliveredAt?: Date | null;
  /** @nullable */
  failedAt?: Date | null;
  /** @nullable */
  failureReason?: string | null;
  metadata?: MessageRecordMetadata;
  /** @nullable */
  actorClerkUserId?: string | null;
  createdAt: Date;
}