# lib/api-zod/src/generated/types/proofPhotoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProofPhotoInputKind } from './proofPhotoInputKind';

export interface ProofPhotoInput {
  kind: ProofPhotoInputKind;
  objectPath: string;
  contentType?: string;
  byteSize?: number;
}